requirements.txt
huggingface_hub==0.22.2
swarmauri[full]==0.4.1
python-dotenv
